+++ dpp-backend/digitalproductpass/src/main/java/org/eclipse/tractusx/digitalproductpass/services/DataTransferService.java
@@ -28,8 +28,8 @@ package org.eclipse.tractusx.digitalproductpass.services;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.eclipse.tractusx.digitalproductpass.config.DtrConfig;
+import org.eclipse.tractusx.digitalproductpass.config.PolicyCheckConfig;
 import org.eclipse.tractusx.digitalproductpass.exceptions.ControllerException;
 import org.eclipse.tractusx.digitalproductpass.exceptions.ServiceException;
 import org.eclipse.tractusx.digitalproductpass.exceptions.ServiceInitializationException;
@@ -41,11 +41,17 @@ import org.eclipse.tractusx.digitalproductpass.models.http.requests.Search;
 import org.eclipse.tractusx.digitalproductpass.models.http.responses.IdResponse;
 import org.eclipse.tractusx.digitalproductpass.models.manager.History;
 import org.eclipse.tractusx.digitalproductpass.models.manager.Status;
-import org.eclipse.tractusx.digitalproductpass.models.negotiation.*;
-import org.eclipse.tractusx.digitalproductpass.models.negotiation.Set;
-import org.eclipse.tractusx.digitalproductpass.models.negotiation.NegotiationTransferResponse;
+import org.eclipse.tractusx.digitalproductpass.models.negotiation.catalog.Catalog;
+import org.eclipse.tractusx.digitalproductpass.models.negotiation.catalog.CatalogRequest;
+import org.eclipse.tractusx.digitalproductpass.models.negotiation.catalog.Dataset;
+import org.eclipse.tractusx.digitalproductpass.models.negotiation.catalog.Offer;
+import org.eclipse.tractusx.digitalproductpass.models.negotiation.policy.Set;
+import org.eclipse.tractusx.digitalproductpass.models.negotiation.response.Negotiation;
+import org.eclipse.tractusx.digitalproductpass.models.negotiation.response.NegotiationTransferResponse;
+import org.eclipse.tractusx.digitalproductpass.models.negotiation.request.NegotiationRequest;
+import org.eclipse.tractusx.digitalproductpass.models.negotiation.request.TransferRequest;
+import org.eclipse.tractusx.digitalproductpass.models.negotiation.response.Transfer;
 import org.eclipse.tractusx.digitalproductpass.models.service.BaseService;
-import org.sonarsource.scanner.api.internal.shaded.minimaljson.Json;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.http.HttpHeaders;
@@ -203,6 +209,39 @@ public class DataTransferService extends BaseService {
         }
     }
     /**
+     * A wrapper function to filter the contract offers and give the valid ones according to the configuration
+     * <p>
+     * @param   catalog
+     *          the {@code Catalog} catalog with the complete offers
+     * @param   policyConfig
+     *          the {@code PolicyCheckConfig} configuration for the policies contained in the contracts
+     * @return  a {@code Map<String, Dataset>} object with the contract offers mapped by id
+     *
+     * @throws  ServiceException
+     *           if unable to get the contract offer for the assetId.
+     */
+    public Map<String, Dataset> getValidContractOffers(Catalog catalog, PolicyCheckConfig policyConfig) throws ServiceException {
+        /*
+         *   This method receives the assetId and looks up for targets with the same name.
+         */
+        try {
+            // Get contracts with the default method
+            Map<String, Dataset> contracts = this.getContractOffers(catalog);
+            if(contracts == null || contracts.keySet().size() == 0){
+                return null;
+            }
+            // If policy check is not enabled return all the contract offers
+            if(!policyConfig.getEnabled()){
+                return contracts;
+            }
+            // If is enabled filter for the valid contract offers
+            return edcUtil.filterValidContracts(contracts, policyConfig);
+        } catch (Exception e) {
+            throw new ServiceException(this.getClass().getName(), e, "It was not possible to get the valid contract Contract Offers");
+        }
+    }
+
+    /**
      * Gets the Contract Offers mapped by contract id
      * <p>
      * @param   catalog
@@ -240,7 +279,6 @@ public class DataTransferService extends BaseService {
             throw new ServiceException(this.getClass().getName(), e, "It was not possible to get Contract Offers for assetId");
         }
     }
-
 
     /**
      * Gets the Contract Offer from the given AssetId in the given provider URL.