+++ CHANGELOG.md
@@ -26,12 +26,24 @@
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
 
+## [released]
+## [v2.3.0] - 06-05-2024
 ### Added
+- SingleApiRequest class for the requested data for the single API.
+- SingleApiConfig class to had configurations related to the single API.
+- ContractService class to move all the processing logic from the API and Contract controllers to this service.
 - Added the following Industry Core changes to the policy and digital twin:
     - Added `manufacturerId` and `digitalTwinType` to the specificAssetIds in digital twin registry
     - Added localIdentifiers to the SerialPart aspect model
+- Updated ApiController with the singleApi POST method.
+- Updated ContractController by creating call methods (create, search, agree and status) without the authentication step to call in the Single API.
+- Updated AuthenticationService by adding the isSingleApiAuthenticated method to authenticate the single API key.
+- Updated application.yaml with the single api configurations.
+- Updated deployment-backend.yaml with the oauth.apiKey.
+- Updated values-int/beta/dev.yaml files with the oauth.apiKey.
 
 ### Updated
 - Refactored workflows where required
@@ -43,7 +55,15 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 - Updated documentation references where required
 - Updated infrastructure guide
 - Updated testdata script to allow EDC constrained policy for the registry
+- Updated ApiController with the singleApi POST method.
+- Updated ContractController by creating call methods (create, search, agree and status) without the authentication step to call in the Single API.
+- Updated AuthenticationService by adding the isSingleApiAuthenticated method to authenticate the single API key.
+- Updated application.yaml with the single api configurations.
+- Updated deployment-backend.yaml with the oauth.apiKey.
+- Updated values-int/beta/dev.yaml files with the oauth.apiKey.
 - Refactored the swagger workflow
+- Updated: as agreed removed BASF from the license and license headers and added CGI instead
+
 
 ### Deleted
 - Deleted unused files/directories/docs/images